requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astroid==3.3.11
attrs==25.3.0
black==25.9.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cryptography==46.0.1
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastuuid==0.12.0
filelock==3.19.1
flake8==7.3.0
Flask==3.1.2
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
litellm==1.77.1
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mistralai==0.4.2
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
odfpy==1.4.1
openai==1.108.1
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pdfminer.six==20250506
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydyf==0.11.0
pyflakes==3.4.0
pylint==3.3.8
pyparsing==3.2.4
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tiktoken==0.11.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
utils==1.0.2
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
xlsxwriter==3.2.9
yarl==1.20.1
zipp==3.23.0
zopfli==0.2.3.post1
